unmask-backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
opencv-python-headless
efficientnet-pytorch